+++ src/curvesimilarities/__init__.py
@@ -1,6 +1,6 @@
 """Curve similarity measures."""
 
-from .averagefrechet import afd, afd_owp, qafd, qafd_owp
+from .averagefrechet import qafd, qafd_owp
 from .dtw import dtw, dtw_owp, sdtw, sdtw_owp
 from .frechet import dfd, fd
 from .integfrechet import ifd, ifd_owp
@@ -14,8 +14,6 @@ __all__ = [
     "sdtw_owp",
     "ifd",
     "ifd_owp",
-    "afd",
-    "afd_owp",
     "qafd",
     "qafd_owp",
 ]